--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/command/application/controller/MypageContoller.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.command.application.controller;
-
-public class MypageContoller {
-}
