--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/command/application/service/BoardService.java
@@ -1,23 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.command.application.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import static org.mybatis.spring.SqlSessionUtils.getSqlSession;
-
-@Service
-@RequiredArgsConstructor
-public class BoardService {
-
-//    private final BoardRepository boardRepository;
-
-//    public BoardService(BoardRepository boardRepository) {
-//        this.boardRepository = boardRepository;
-//    }
-
-//    public List<Board> selectBoardList(BoardListDTO boardListDTO) {
-//
-////        List<BoardListDTO> boardList = boardRepository.selectBoardList(boardListDTO.getBoardNo());
-//    }
-//}
-}
