--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/command/infra/repository/Repository.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.command.infra.repository;
-
-public class Repository {
-}
