--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/command/infra/service/Service.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.command.infra.service;
-
-public class Service {
-}
