--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/query/application/Application.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.query.application;
-
-public class Application {
-}
