--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/query/domain/Domian.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.query.domain;
-
-public class Domian {
-}
