--- src/main/java/com/ohgiraffers/goonthatbackend/metamate/board/query/infra/Infra.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.goonthatbackend.metamate.board.query.infra;
-
-public class Infra {
-}
