+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/command/application/controller/MypageContoller.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.command.application.controller;
+
+public class MypageContoller {
+}
