+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/command/application/service/FreeBoardService.java
@@ -1,0 +1,12 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.command.application.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import static org.mybatis.spring.SqlSessionUtils.getSqlSession;
+
+@Service
+@RequiredArgsConstructor
+public class FreeBoardService {
+
+}
