+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/command/domain/repository/FreeBoardRepository.java
@@ -1,0 +1,8 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.command.domain.repository;
+import com.ohgiraffers.goonthatbackend.metamate.freeboard.command.domain.aggregate.entity.FreeBoard;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FreeBoardRepository extends JpaRepository<FreeBoard,Long> {
+
+
+}
