+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/command/infra/repository/Repository.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.command.infra.repository;
+
+public class Repository {
+}
