+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/command/infra/service/Service.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.command.infra.service;
+
+public class Service {
+}
