+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/query/application/Application.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.query.application;
+
+public class Application {
+}
