+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/query/domain/Domian.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.query.domain;
+
+public class Domian {
+}
