+++ src/main/java/com/ohgiraffers/goonthatbackend/metamate/freeboard/query/infra/Infra.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.goonthatbackend.metamate.freeboard.query.infra;
+
+public class Infra {
+}
